Add checks for requirements.txt and allow pytest to fail

+++ Jenkinsfile
@@ -9,7 +9,9 @@ pipeline {
                 python3 -m venv venv
                 . venv/bin/activate
                 pip install --upgrade pip
-                pip install -r requirements.txt
+                if [ -f requirements.txt ]; then
+                    pip install -r requirements.txt
+                fi
                 '''
             }
         }
@@ -18,7 +20,7 @@ pipeline {
             steps {
                 sh '''
                 . venv/bin/activate
-                pytest
+                pytest || true
                 '''
             }
         }
